Changes never committed

+++ lib/ews-landslide-library/SoilMoisture/SoilMoisture.cpp
@@ -2,6 +2,6 @@
 
 SoilMoisture::SoilMoisture(uint8_t pin) : _pin(pin) { }
 
-uint8_t SoilMoisture::get_soil_moisture_reading() {
+uint16_t SoilMoisture::get_soil_moisture_reading() {
     return analogRead(this->_pin);
 }
